+++ tests/ai_safety/guardrails/test_guardrails.py
@@ -6,7 +6,7 @@ from ocp_resources.custom_resource_definition import CustomResourceDefinition
 from simple_logger.logger import get_logger
 from timeout_sampler import retry
 
-from tests.model_explainability.guardrails.constants import (
+from tests.ai_safety.guardrails.constants import (
     AUTOCONFIG_DETECTOR_LABEL,
     PII_INPUT_DETECTION_PROMPT,
     PII_OUTPUT_DETECTION_PROMPT,
@@ -20,7 +20,7 @@ from tests.model_explainability.guardrails.constants import (
     STANDALONE_DETECTION_ENDPOINT,
     AUTOCONFIG_GATEWAY_ENDPOINT,
 )
-from tests.model_explainability.guardrails.utils import (
+from tests.ai_safety.guardrails.utils import (
     create_detector_config,
     verify_health_info_response,
     send_and_verify_unsuitable_input_detection,
@@ -28,7 +28,7 @@ from tests.model_explainability.guardrails.utils import (
     send_and_verify_negative_detection,
     send_and_verify_standalone_detection,
 )
-from tests.model_explainability.utils import validate_tai_component_images
+from tests.ai_safety.utils import validate_tai_component_images
 from utilities.constants import (
     LLM_D_CHAT_GENERATION_CONFIG,
     BUILTIN_DETECTOR_CONFIG,
@@ -41,7 +41,7 @@ LOGGER = get_logger(name=__name__)
 
 
 @pytest.mark.smoke
-@pytest.mark.model_explainability
+@pytest.mark.ai_safety
 def test_guardrailsorchestrator_crd_exists(
     admin_client: DynamicClient,
 ) -> None:
