+++ tests/ai_safety/trustyai_service/conftest.py
@@ -25,7 +25,7 @@ from ocp_resources.serving_runtime import ServingRuntime
 from ocp_resources.trustyai_service import TrustyAIService
 from pytest_testconfig import py_config
 
-from tests.model_explainability.trustyai_service.constants import (
+from tests.ai_safety.trustyai_service.constants import (
     GAUSSIAN_CREDIT_MODEL,
     GAUSSIAN_CREDIT_MODEL_RESOURCES,
     GAUSSIAN_CREDIT_MODEL_STORAGE_PATH,
@@ -40,10 +40,10 @@ from tests.model_explainability.trustyai_service.constants import (
     TAI_PVC_STORAGE_CONFIG,
     XGBOOST,
 )
-from tests.model_explainability.trustyai_service.trustyai_service_utils import (
+from tests.ai_safety.trustyai_service.trustyai_service_utils import (
     wait_for_isvc_deployment_registered_by_trustyai_service,
 )
-from tests.model_explainability.trustyai_service.utils import (
+from tests.ai_safety.trustyai_service.utils import (
     create_isvc_getter_role,
     create_isvc_getter_role_binding,
     create_isvc_getter_service_account,
