+++ tests/ai_safety/trustyai_service/drift/test_drift.py
@@ -2,8 +2,8 @@ from functools import partial
 
 import pytest
 
-from tests.model_explainability.trustyai_service.constants import DRIFT_BASE_DATA_PATH
-from tests.model_explainability.trustyai_service.trustyai_service_utils import (
+from tests.ai_safety.trustyai_service.constants import DRIFT_BASE_DATA_PATH
+from tests.ai_safety.trustyai_service.trustyai_service_utils import (
     TrustyAIServiceMetrics,
     send_inferences_and_verify_trustyai_service_registered,
     verify_trustyai_service_metric_delete_request,
