+++ tests/ai_safety/trustyai_service/fairness/conftest.py
@@ -10,7 +10,7 @@ from ocp_resources.secret import Secret
 from ocp_resources.service import Service
 from ocp_resources.serving_runtime import ServingRuntime
 
-from tests.model_explainability.trustyai_service.trustyai_service_utils import (
+from tests.ai_safety.trustyai_service.trustyai_service_utils import (
     wait_for_isvc_deployment_registered_by_trustyai_service,
 )
 from utilities.constants import MinIo, ModelFormat, KServeDeploymentType, RuntimeTemplates
