+++ tests/ai_safety/trustyai_service/service/conftest.py
@@ -9,12 +9,12 @@ from ocp_resources.secret import Secret
 from ocp_resources.trustyai_service import TrustyAIService
 
 
-from tests.model_explainability.trustyai_service.constants import (
+from tests.ai_safety.trustyai_service.constants import (
     TAI_METRICS_CONFIG,
     TAI_DB_STORAGE_CONFIG,
 )
 from utilities.constants import TRUSTYAI_SERVICE_NAME
-from tests.model_explainability.trustyai_service.utils import (
+from tests.ai_safety.trustyai_service.utils import (
     create_trustyai_service,
 )
